% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simAnneal.R
\name{sim.geno.data}
\alias{sim.geno.data}
\title{Generates a simulated aannealing function and computes coverage required}
\usage{
sim.geno.data(data = NULL, plot = TRUE)
}
\arguments{
\item{data}{a dataframe with outcome and predictor variables (e.g. genetic variants).}

\item{plot}{a boolean, TRUE to plot the function and FALSE otherwise}
}
\value{
a vector and optionally a plot
\code{numVarRetained} a vector of integrers with length = number of itrations required
}
\description{
To be written.
}
\author{
Barry C.
}

